The main change for this routine:
DiskVoteBlocks had some code that was supposed to speed up retrieval when retrieval was in order.  In practice, the speed-up could give the wrong value.

I've added a test specific to the error in the speed-up, as well as a few extra tests (just because.)

IMPORTANT NOTE: I have changed V3TestUtils so that "makeVoteBlockList" returns a vote block list with different-sized entries.  This change is important; not having it will obscure many tests.


git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@7231 4f837ed2-42f5-46e7-a7a5-fa17313484d4