The ListHoldings servlet can now be disabled by adding org.lockss.config.ui.showHoldingsServlet = false to lockss.opt.
src/org/lockss/config/ConfigManager.java
Added a new parameter.
src/org/lockss/servlet/AdminServletManager.java
Used the parameter to dictate whether the Holdings servlet should be created and shown in the UI menu.
If the parameter is not specified or is something other than false, the servlet is shown. It must be explicitly disabled.


git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@10561 4f837ed2-42f5-46e7-a7a5-fa17313484d4