Refactor UrlTallier to move getVoterUrlTally, getPollerUrlTally. and getRepairUrlTally to V3Poller. This moves UrlTallier closer to being sharable with the symmetric polling code.
Add explicit addBlockTally/getBlockTally routines to VoteBlockTallier,
and use those rather than use the general addTally. This allows the
getVoterUrlTally &c routines above to return a VoteBlockTallier and
later recover the BlockTally.

Unit tests are largely unchanged, to show that this was largely just a
code-rearrangement. The tests should be refactored based on this
change, and improved dependency injection should allow more tests.


git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@28099 4f837ed2-42f5-46e7-a7a5-fa17313484d4