Split new class VoteBlocksCoordinator from UrlTallier. [VoteBlocksCoordinator and TestVoteBlocksCoordinator were submitted in a previous commit.]
The coordination of a collection of VoteBlockIterator instances has
been factored out and is in the new class. This makes testing easier
and clarifies the coordination effort out of the context of
voting. The UrlTallier is responsible for the V3Poller's use of the
VoteBlocksCoordinator.

The VoteBlocksCoordinator is also now used by the V3Voter in
symmetric polls.

Split testing of VoteBlocksCoordinator off from testing of UrlTallier.


git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@28421 4f837ed2-42f5-46e7-a7a5-fa17313484d4