Roundup issue 5036 (https://roundup.lockss.org/lockss/issue5036): Delete AUs behavior is confusing when subscription manager in use.
Added code to add the AU to the UNCONFIGURED_AU table when an AU is deleted; also added code to remove the AU from the UNCONFIGURED_AU table when an AU is added. This will keep every AU either in the AU table or the UNCONFIGURED_AU table.

Added code to unsubscribe an AU when it is deleted. This will prevent the subscription manager to add the AU back on the next configuration change.


git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@35341 4f837ed2-42f5-46e7-a7a5-fa17313484d4