Roundup issue 5076 (https://roundup.lockss.org/lockss/issue5076): Allow multiple proprietary identifiers for any given publication.
To record the various proprietary identifiers for any given publication, a new PROPRIETARY_ID table has been created in the database. It contains the following columns:

* MD_ITEM_SEQ - A foreign key pointing to the primary key of the publication in the MD_ITEM table.

* PROPRIETARY_ID - The text of the proprietary identifier, like the existing PUBLICATION.PUBLICATION_ID column.

Each row has the constraint to be unique.

The existing PUBLICATION.PUBLICATION_ID column has been removed.

Also, TdbTitle and BibliographicItem have been modified to provide the array of proprietary identifiers, instead of a single one.

In the COUNTER reports and in the subscription backup file, all the proprietary identifiers are included, separated by commas.


git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@37399 4f837ed2-42f5-46e7-a7a5-fa17313484d4