Roundup issue 5145 (https://roundup.lockss.org/lockss/issue5145): Allow Subscribe All and Subscribe Publisher.
Modified so that processing of AUs that appear as new in any configuration change set happens always in a separate thread, not just for the first change set after the daemon starts, as until now. Added the handling of the "Subscribe All" flag when determining whether a new AU needs to be configured. Placed the configuration of AUs under a rate limiter.

git-svn-id: svn://svn.code.sf.net/p/lockss/svn/lockss-daemon/trunk@42410 4f837ed2-42f5-46e7-a7a5-fa17313484d4